rewrite tests in Deno style
