+++ deno.json
@@ -15,6 +15,7 @@
     "src/": "./src/",
     "deps": "./src/deps.ts",
     "consts": "./src/constants.ts",
-    "test-deps": "./src/tests/deps.ts"
+    "test-deps": "./src/tests/deps.ts",
+    "get-project-root": "./_here.ts"
   }
 }
